chore: extract ChatMLSchema to shared package
