+++ web/src/components/ui/MarkdownViewer.tsx
@@ -26,7 +26,7 @@ import {
   type OpenAIOutputAudioType,
   isOpenAITextContentPart,
   isOpenAIImageContentPart,
-} from "@/src/components/schemas/ChatMlSchema";
+} from "@langfuse/shared";
 import { type z } from "zod/v4";
 import { ResizableImage } from "@/src/components/ui/resizable-image";
 import { LangfuseMediaView } from "@/src/components/ui/LangfuseMediaView";
